Add Emotions plot to app
A table to visualise emotive words was added to the output options as a
conditonal panel. Also some changes were made to the code to make for
better readability, notably the creation of objects in renderPlot() before
the execution of conditional statements that are dependend on the
conditional input values.