Implement back-end storage
I just focused on setting up the database for the tweets and made a few associated changes to the code.

I had some doubts whether the repo or the app directory is the root and I could not confirm it at the time experimentally, as the tweets on nesreanigeria returned zero at the time of writing the code. I didn't want to try other search terms so as not to mess up the data that had already been stored. So that has to be worked on.