Display most positive/negative tweets
Functionality for the display of tweets from both extremes of the emotive spectrum was incorporated in the app. This is to enable effective guaging of the value of the so-called Emotions Plot as well as to show whether there is a concentration of positive or negative words in a particular tweet.