Create a session log
A text file that records a few simple events from the Shiny session was created. Significantly, record of quantitative data on tweets downloaded and stored with the search term "nesreanigeria" (which is hard-coded into the app, since that is what interests us the most at NESREA). The actual log file will not be committed.