Enable bookmarking
The bookmarkable state is now included in the app. When a user clicks on the "Bookmark" button, a link that can be shared with any other person is provided. When this link is run in a web browser, the state of the app at the time of the creation of the bookmark is displayed. Pretty cool!