Fix incorrect annotation of graph
What was previously used as the x-axis label was totally incorrect. First the code was corrected for proper display of the dates and the entire label changed to a title (via ggtitle()). Also, the new x-axis label now correctly named "Date", cos that's what it is. I feel like slapping myself.