Fix bug that prevented knitting
In line 158, rather than substituting the "-" with a non-character, this was replaced with an NA object. The file now knits appropriately but the snag is that the NAs are reflected in the table. This cam be removed later, but I'm procrastinating for now for lack of time and to allow for the release of the report TODAY, as scheduled.