Obtain relevant Facebook insights
I wrote a simple function out of the existing code (line 10 - 26). This function will allow us to run a number of options for downloading insights from Facebook as seen in lines 28 - 35. This greatly simlifies the process.

In line 7 I added code that sources the authentication details that allow access to the Facebook API. Members of the team should ask for this file and I will send it via email. Add it to the NESREA_social main repo directory so that the file can find it and it will also get added to .gitignore and will not be added to your own commits.