Skips wordcloud when data are insuffient
The document was not knitting and it was discovered that the length of the list of "signed" emotional valences i.e. -1, 0 and +1 for "negative, neutral and positive" did not match, because no neutral tweets were downloaded in the first place. This happened basically because there were relatively few data available for analysis.

This change enables the document to skip trying to render the wordcloud whenever there aren't enough data, thus avoiding the error message and allowing knitting to proceed beyond this point.