Redirects link for feedback
Rather than emails to maintainer(s), the link for "Feedback/Complaints"
has been recoined and now points to the GitHub Issues page for this
repository. If a user clicks on the link, he/she will immediately open a
"New Issue". Also, a few minor aesthetic changes were made to the user interface.