Includes option for downloads
Before now, the report either downloaded data by default or didn't do so at all. So we were faced with a dilemna where, on one hand one would not want to download every single time the document is knitted (for performance reasons) and on the other run into an error whenever there was insufficient new data (the report is to be 'current') and would be forced to open another script or run code in the console to do so (thereby wasting time and energy). Programmers are indeed a lazy bunch...