Introduces conditional R script sourcing
The source() function has been wrapped in a user-defined function that throws a warning in the event that the sourced file does not exist in the working directory.

Also, the repeated calls to library() has been improved with the use of the functional lapply(), which works through a character vector of the package names.