Fixes bug responsible for knitting failure
The main problem was in the current line 262, where sapply() was oversimplifying, and thus distorting, the table of positive/negative words. This part of the code worked hitherto, so I guess R Core Team must have introduced or modified that function argument.

Other minor changes were mostly general improvements to the code.